Add visual editor gizmo skill and save-editor extension
