+++ .gitignore
@@ -3,3 +3,4 @@ dist/
 *.tsbuildinfo
 .opendcl/
 .pi/
+.DS_Store
