--- gizmo-prototype-v2/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "gizmo-prototype-v2",
-  "version": "1.0.0",
-  "description": "Gizmo prototype with ray-plane intersection drag",
-  "scripts": {
-    "start": "sdk-commands start",
-    "build": "sdk-commands build"
-  },
-  "devDependencies": {
-    "@dcl/js-runtime": "7.20.4",
-    "@dcl/sdk": "latest"
-  },
-  "engines": {
-    "node": ">=16.0.0"
-  }
-}
