--- gizmo-prototype-v2/scene.json
@@ -1,21 +1,0 @@
-{
-  "ecs7": true,
-  "runtimeVersion": "7",
-  "display": {
-    "title": "Gizmo Prototype v2",
-    "description": "Ray-plane intersection drag approach"
-  },
-  "scene": {
-    "parcels": ["0,0", "1,0", "0,1", "1,1"],
-    "base": "0,0"
-  },
-  "main": "bin/index.js",
-  "spawnPoints": [
-    {
-      "name": "spawn1",
-      "default": true,
-      "position": { "x": [1, 3], "y": [0, 0], "z": [1, 3] },
-      "cameraTarget": { "x": 16, "y": 1, "z": 16 }
-    }
-  ]
-}
