--- gizmo-prototype-v2/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "strict": true,
-    "baseUrl": "."
-  },
-  "include": ["src/**/*.ts", "src/**/*.tsx"],
-  "extends": "@dcl/sdk/types/tsconfig.ecs7.json"
-}
